надеюсь, что это всё

+++ task_04/src/order_statistics.hpp
@@ -3,4 +3,4 @@
 #include <stdexcept>
 #include <vector>
 
-int GetOrderStatistics(const std::vector<int>& data, size_t n);
+int GetOrderStatistics(const std::vector<int>& data, int n);
